Fix Operation filtering when multiple service models exists on the classpath
